Restructure split pages from views, /graphql -> /queries
