+++ client/src/pages/README.md
@@ -1,0 +1,3 @@
+Pages handle bringing static and dynamic components together.
+
+Pages shouldn't handle fetching data directly. If a component requires dynamic data it should be included as a view.
