+++ client/src/pages/index.js
@@ -1,0 +1,2 @@
+export {default as HomePage} from './home-page';
+export {default as DetailPage} from './detail-page';
