+++ client/src/queries/categories-query.js
