+++ client/src/queries/detail-query.js
