+++ client/src/views/README.md
@@ -1,0 +1,1 @@
+Views wrap components that rely on fetched data.
